dedicated actions for manipulating the packet metadata
The edit action is no longer a universal data transfer. On the other hand
the code is less abstract now, and simplicity is more important.
The biggest winners are the stateful objects (SeqGen, SeqRcvy, POF), which
now work with the packet metadata directly without using the convoluted
value transfer framework.