rework ConfObject into a base class for stateful pipeline objects
ConfObject was originally meant to be a temporary storage for the config
parser, but it turned out to be a permanent one. Let's do it properly.

I think the interfaces are cleaner this way, especially for future
additions.

A lot of changes in a lot of files, but mostly just type adjustments.