log: stop using GNU-specific features
I always try to avoid non-standard API. Now we no longer use strndupa, and
we use the standard version of strerror_r. Enabled _FORTIFY_SOURCE in the
Makefile to ensure that we get a warning if we got the GNU version of
strerror_r (on some systems _FORTIFY_SOURCE is enabled by default).

Also forbid variable-length arrays and do some include cleanup.