makefile: properly redefine _FORTIFY_SOURCE
Ubuntu defines this implicitly to a number that they keep changing, and if
it's not 3 we would get an error.

BTW some folks (e.g. Wine) are arguing that _FORTIFY_SOURCE has some false
positives and doesn't catch real errors... We started using it to make sure
we don't get the GNU version of strerror_r().