mask: React to mask/unmask request in OAM endpoint
Mask signals exclusively handled by auto-generated MIPs. The signal
itself link-local, so the target always a next-hop. The pre-Eliminiation
MIP traps the signal, update the last heartbeat timestamp then drop the
OAM packet.

The signal handling process looks like the following:
1. Elimination pre-MIP receive the signal
2. Update the last heartbeat timestamp, drop the signal
3. Notify the post-MIP thread about the new mask/unmask signal
4. The post-MIP thread wakes up, check all pre-MIP's of the Elim object
5. Count how many masked, and update the latent error paths accordingly

There are two interesting cases:
A. When all pre-MIP masked, the post-MIP start a new mask session. This
   is necessary, since subsequent Eliminations in the recovery graph
   notified that way to update their path count.
B. If the post-MIP in state A. and one pre-MIP unmasked, the mask
   session terminated and an unmask signal generated to notify the
   subsequent nodes in the recovery graph