log: packettrace: Move packet trace logging to delete_packet
Move packet trace logging to the delete_packet function.
The rationale behind this is the fact that we release
packets at the end of their lifecycle.
Since we have done this properly, the fate of the packet is
already encoded in the packet's log buffer.

This also fixes the silent elimination drops as well, and
starting with this, we log packet sequences at elimination.