doc: Simplify and consolidate scenario2 (TSNoDetNet)
Getting started scenario 2 was highly fragmented:
- base files (nxp[12].ini)
- ipv6 version (6nxp[12].ini)
- priority version (nxp[12]_pcp.ini)
- commented base files (commented_nxp[12].ini)

We drop IPv6 for now, since it does not tell much about the story.
We can re-introduce it later on other getting started perhaps for the
IPoDetNet scenarios.

The priority version merged into the base scenario, so we have another
stream with PCP=6 priority tag, matching on another stream.

The commented version dropped too, since everything explained in the
README.md (and that should be improved if not).