oam: fix: Set OAM packet length
The length of a fresh OAM packet was uninitialized. This usually not
problem, unless the following action pipeline:

... mep-start foo 3, mip bar 3, send nni

There MIP bar reinterpret the header structure and there is a part when
it calculate the payload:

hdr[2].len = p->len - 4 - 8;

There p->len=0 since we still have the same struct Packet created by
MEP foo and the subtraction resulting in an underflow, cause sendmsg
fail with "Message too big"

The fix is calculate the p->len right after the MEP start filled the
request.