recovery: Fix use-after-free in decrement_ticks
In temporary OAM recovery case, ticks=0 means stop thread and free
recovery object. However until this finished, decrement_ticks running in
the reset_thread on the already (or partially) free'd recovery object.

This results in a huge UAF spam in valgrind in case of a ping any
(annoying in case of pinging 1 MP as well).