oam: Fix MIP stats type error
With -o the value type for key "target" changed from string to object
resulting type mismatch at reply processing: "No target in reply." This
is because we tried to get key as string which is only OK if -o not
given.

For now workaround the issue with:
* leave "target" as a string (value is the target MIP's ID)
* add "target_info" key with JSON obj as value (with the counters)
* harmonize the new "source" key by renaming it to "source_info"

In the future we might came up with better OAM JSON messages, therefore
this might changes. Right now documentation is modified accordingly to
let the user aware of these JSON records.