notification: Insert last notification segment
When the length of the notification reaches MAX_NOTIFICATION_LEN we do
not insert the actual notification to the message and after that we
continue the iteration from the next pull source: we lost the notification.

Now we initialize the new packet with the actual notification which does
not fit into the previous packet. We set total_length accordingly.