CI: attempt to fix double pipeline for this branch
There is a pipeline for the commit and one for the merge request. Let's try
to suppress the commit pipeline if there is a merge request. Unfortunately
it's impossible to test the CI config without committing it.

It would also be nice if we could avoid running a pipeline if the commit
only touches the documentation...