test: pof: Workaround for test flakyness
If seq 2 received before packet seq 1 our test not work properly. This
is because we always check if seq 1 is the 1 packet, but POF cannot "see
the future" so its normal operation for the TakeAny status POF to let
seq 2 first then withhold seq 1 in the delaybuffer.

Fix: if seq 1 is not the first, then check if it is delayed for max POF
delay (global paramter of the test SHOULD be consistent with the
config). If is, we know its safe to omit lets say the first 10 packets
from rigorous ordering test.