oam: fix node id in d-ACH
It's a 20 bit field, so fill it as such. Note that currently in OAM our
node id is the last 16 bits of the IP address of the default return
interface.

I noticed this while debugging a flaky check in the seq_rcvy_oam test. I
don't yet know why the test is flaky.