oam: rework the actions, hook up the new message receiver
This fixes rlist and ruins the mask signalling. Fair trade.

Note: now the list command shows the stream name where the MP is, not the
name at the beginning of the pipeline. This is the main reason why the
OAM AutoMIP test fails. The other reason is that I've changed the format
of the printout :)