ci: Resolve gitlab IP manually
This is an ugly workaround on the DNS resolution.
There might be something wrong with NSCD daemon, first of all the daemon
not installed and not running in the CI containers. Even if it running,
the first lookup takes ~5 sec, enough for timeout.

Since I cant found a clean way to use NSCD properly or force CI
containers to /etc/resolv.conf what we can do is lookup manually.
The result then appended to hosts and no resolution needed.