conf_actions: remember the header, field names in the edit assignments
Knowing these names in addition to the indexes and offsets will help the
pipeline analyzer find where a certain field has been written.

Also, DEL now keeps the descriptor for the deleted header for similar
reasons.

For now this is only used in the debug logging.

TODO I had to add code for filling the header/field names in way too many
places, which means we need more helpers for constructing assignments