conf_actions: analyze the action list to find addressing for oam injection
Now we can supply the MP the address necessary for message injection if
- there was an EDIT that wrote the field
- there was a match for the field

We can also tell the MP if the necessary field will be written later.

Change in behavior: we used to bail if the address field wasn't written,
from now on the MP will simply not be able to inject messages. There are a
couple of warnings about this.
This should fix #6

Currently the addresses are handed over to the MP, but it doesn't yet do
anything with it. Also, everything is untested.