seq_recov: fix latent error paths
The admin_latent_error_paths was 0 unless frerSeqRcvyLatentErrorPaths was
specified in the config.

Moved latent_error_paths out of RecoveryDiagnosticConf, this is not a
configured parameter, but the current path count (affected by masking).

BTW the admin_latent_error_paths could be automatically derived from the
number of Elimination actions referring to this object.