oam: reimplement the path masking
The original implementation was ruined by !24 and its remains have been
commented out since.

This new implementation should be easier to read, more flexible, and it's
able to process the unmask signal.

I'm not entirely sure about the mask signal regeneration, though, it's hard
to test with the current test network.