notification: add a randomly generated constant session id
This replaces 420559b7 where I used the source port for the session id in
the de-duplication of messages with the same host and seq to mark the ones
coming from a newly started r2dtwo not duplicate. However, when r2dtwo is
configured to send the notifications on multiple interfaces for redundancy,
those will have different source port, but they must be de-duplicated.
Thus, the source port is not a good session id.

The probability of getting the same session id twice should be around one
in two billion.