oam maintenance: fix crash on shutdown
Now OAM MPs remember all the injection points, and use one from the list.
Deleting a pipeline now notifies OAM about it, and any MP that uses it to
inject will remove it from its list. This fixes a crash on shutdown when
a background ping session wanted to send after the pipeline was deleted.

Note that previously MP used the first registered injection point, now it
uses the last one. The tests are updated for this.